Refactor move preload scripts
